Use 8 bit integers
* src/components/games/games.h (titctac_move): Change xp and yp to
  uint8_t pointers.

* src/components/games/games.c (send_move): Change x and y to uint8_t.

+++ src/components/games/games.c
@@ -95,7 +95,7 @@ void game_register(uint8_t snode)
 	lownet_send(&pkt);
 }
 
-int send_move(int x, int y)
+int send_move(uint8_t x, uint8_t y)
 {
 	lownet_frame_t pkt;
 	game_action_t *ga = (game_action_t*) pkt.payload;
